Move ToServices translation into operator
The pkg/policy/groups package was polluted by Kubernetes specific CNP
handling.
Move all of this into operator/ so it can be reused for services
translation.
