+++ operator/cnp_event.go
@@ -16,20 +16,25 @@ package main
 
 import (
 	"context"
+	"fmt"
+	"sync"
 	"time"
 
 	"github.com/cilium/cilium/pkg/controller"
 	"github.com/cilium/cilium/pkg/k8s"
 	cilium_v2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
+	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/metrics"
-	"github.com/cilium/cilium/pkg/policy/groups"
 
+	"github.com/sirupsen/logrus"
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/tools/cache"
 )
+
+const maxConcurrentUpdates = 4
 
 func init() {
 	runtime.ErrorHandlers = []func(error){
@@ -48,7 +53,20 @@ func enableCNPWatcher() error {
 			AddFunc: func(obj interface{}) {
 				metrics.EventTSK8s.SetToCurrentTime()
 				if cnp := k8s.CopyObjToV2CNP(obj); cnp != nil {
-					groups.AddDerivativeCNPIfNeeded(cnp.CiliumNetworkPolicy)
+					if !cnp.RequiresDerivative() {
+						log.WithFields(logrus.Fields{
+							logfields.CiliumNetworkPolicyName: cnp.ObjectMeta.Name,
+							logfields.K8sNamespace:            cnp.ObjectMeta.Namespace,
+						}).Debug("CNP does not have derivative policies, skipped")
+						return
+					}
+
+					controllerManager.UpdateController(fmt.Sprintf("add-derivative-cnp-%s", cnp.ObjectMeta.Name),
+						controller.ControllerParams{
+							DoFunc: func(ctx context.Context) error {
+								return addDerivativeCNP(cnp.CiliumNetworkPolicy)
+							},
+						})
 				}
 			},
 			UpdateFunc: func(oldObj, newObj interface{}) {
@@ -59,7 +77,30 @@ func enableCNPWatcher() error {
 							return
 						}
 
-						groups.UpdateDerivativeCNPIfNeeded(newCNP.CiliumNetworkPolicy, oldCNP.CiliumNetworkPolicy)
+						if !newCNP.RequiresDerivative() {
+							if oldCNP.RequiresDerivative() {
+								log.WithFields(logrus.Fields{
+									logfields.CiliumNetworkPolicyName: newCNP.ObjectMeta.Name,
+									logfields.K8sNamespace:            newCNP.ObjectMeta.Namespace,
+								}).Info("New CNP does not have derivative policy, but old had. Deleted old policies")
+
+								controllerManager.UpdateController(fmt.Sprintf("delete-derivatve-cnp-%s", oldCNP.ObjectMeta.Name),
+									controller.ControllerParams{
+										DoFunc: func(ctx context.Context) error {
+											return DeleteDerivativeCNP(oldCNP.CiliumNetworkPolicy)
+										},
+									})
+							}
+
+							return
+						}
+
+						controllerManager.UpdateController(fmt.Sprintf("CNP-Derivative-update-%s", newCNP.ObjectMeta.Name),
+							controller.ControllerParams{
+								DoFunc: func(ctx context.Context) error {
+									return addDerivativeCNP(newCNP.CiliumNetworkPolicy)
+								},
+							})
 					}
 				}
 			},
@@ -81,7 +122,7 @@ func enableCNPWatcher() error {
 				}
 				// The derivative policy will be deleted by the parent but need
 				// to delete the cnp from the pooling.
-				groups.DeleteDerivativeFromCache(cnp.CiliumNetworkPolicy)
+				DeleteDerivativeFromCache(cnp.CiliumNetworkPolicy)
 			},
 		},
 		k8s.ConvertToCNP,
@@ -91,11 +132,50 @@ func enableCNPWatcher() error {
 	controller.NewManager().UpdateController("cnp-to-groups",
 		controller.ControllerParams{
 			DoFunc: func(ctx context.Context) error {
-				groups.UpdateCNPInformation()
+				// Retrieves all the CNP that has currently a
+				// derivative policy and creates the new
+				// derivatives policies with the latest
+				// information from providers.  To avoid issues
+				// with rate-limiting this function will
+				// execute the addDerivative function with a
+				// max number of concurrent calls, defined on
+				// maxConcurrentUpdates.
+				cnpToUpdate := cnpCache.GetAllCNP()
+				sem := make(chan bool, maxConcurrentUpdates)
+				for _, cnp := range cnpToUpdate {
+					sem <- true
+					go func(cnp *cilium_v2.CiliumNetworkPolicy) {
+						defer func() { <-sem }()
+						addDerivativeCNP(cnp)
+					}(cnp)
+				}
 				return nil
 			},
 			RunInterval: 5 * time.Minute,
 		})
 
 	return nil
+}
+
+var cnpCache = cnpCacheMap{}
+
+type cnpCacheMap struct {
+	sync.Map
+}
+
+func (cnpCache *cnpCacheMap) UpdateCNP(cnp *cilium_v2.CiliumNetworkPolicy) {
+	cnpCache.Store(cnp.ObjectMeta.UID, cnp)
+}
+
+func (cnpCache *cnpCacheMap) DeleteCNP(cnp *cilium_v2.CiliumNetworkPolicy) {
+	cnpCache.Delete(cnp.ObjectMeta.UID)
+}
+
+func (cnpCache *cnpCacheMap) GetAllCNP() []*cilium_v2.CiliumNetworkPolicy {
+	result := []*cilium_v2.CiliumNetworkPolicy{}
+	cnpCache.Range(func(k, v interface{}) bool {
+		result = append(result, v.(*cilium_v2.CiliumNetworkPolicy))
+		return true
+	})
+	return result
 }
